Create 'svnadmin archive' subcommand.
Just as what used to happen with 'db_recover', a lot of users are now
running the wrong version of 'db_archive'.  So we're wrapping this BDB
function in svnadmin, so it will always Just Work (correct library
version, no need to remember -a or -h flags.)

IMO, 'svnadmin archive' and 'svnadmin recover' are the only two
BDB maintenance commands that svn admins really need to know, and
hopefully svnadmin has made them easier to use.

* libsvn_fs/fs.c, svn_fs.h (svn_fs_berkeley_archive): new wrapper
  around the bdb ENV->archive() function.

* svnadmin/main.c (svn_opt_subcommand_t, cmd_table):  add 'archive' subcommand.
  (subcommand_archive):   new function.

* project_faq.html:  mention 'svnadmin archive' in FAQ.

* book/ch05.xml (Berkeley DB Utilities):  show example of 'svnadmin archive'.

* book/ch08.xml (svnadmin): document new subcommand.